+++ docs/release-notes.md
@@ -13,6 +13,7 @@
   * Added support for `ISemanticVersion` in JSON models.
   * Fixed deadlock in rare cases when injecting a file with an asset loader.
   * Fixed unhelpful error when a mod exposes a non-public API.
+  * Fixed input events being raised for keyboard buttons when a textbox is receiving input.
 
 * For SMAPI developers:
   * Overhauled mod DB format to be more concise, reduce the memory footprint, and support versioning/defaulting more fields.